chore: combine test editor pages
